Extract test cases regarding pareto front to test_multi_objective.py
