+++ tests/study_tests/test_study.py
@@ -9,7 +9,6 @@ from typing import Callable
 from typing import Dict
 from typing import List
 from typing import Optional
-from typing import Tuple
 from unittest.mock import Mock  # NOQA
 from unittest.mock import patch
 import uuid
@@ -822,64 +821,12 @@ def test_optimize_with_multi_objectives(n_objectives: int) -> None:
         assert len(trial.values) == n_objectives
 
 
-def test_pareto_front_2d() -> None:
-    def _trial_to_values(t: FrozenTrial) -> Tuple[float, ...]:
-        assert t.values is not None
-        return tuple(t.values)
-
+def test_best_trials() -> None:
     study = create_study(directions=["minimize", "maximize"])
-    assert {_trial_to_values(t) for t in study.best_trials} == set()
-
     study.optimize(lambda t: [2, 2], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(2, 2)}
-
     study.optimize(lambda t: [1, 1], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 1), (2, 2)}
-
     study.optimize(lambda t: [3, 1], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 1), (2, 2)}
-
-    study.optimize(lambda t: [3, 2], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 1), (2, 2)}
-
-    study.optimize(lambda t: [1, 3], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 3)}
-    assert len(study.best_trials) == 1
-
-    study.optimize(lambda t: [1, 3], n_trials=1)  # The trial result is the same as the above one.
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 3)}
-    assert len(study.best_trials) == 2
-
-
-def test_pareto_front_3d() -> None:
-    def _trial_to_values(t: FrozenTrial) -> Tuple[float, ...]:
-        assert t.values is not None
-        return tuple(t.values)
-
-    study = create_study(directions=["minimize", "maximize", "minimize"])
-    assert {_trial_to_values(t) for t in study.best_trials} == set()
-
-    study.optimize(lambda t: [2, 2, 2], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(2, 2, 2)}
-
-    study.optimize(lambda t: [1, 1, 1], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 1, 1), (2, 2, 2)}
-
-    study.optimize(lambda t: [3, 1, 3], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 1, 1), (2, 2, 2)}
-
-    study.optimize(lambda t: [3, 2, 3], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 1, 1), (2, 2, 2)}
-
-    study.optimize(lambda t: [1, 3, 1], n_trials=1)
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 3, 1)}
-    assert len(study.best_trials) == 1
-
-    study.optimize(
-        lambda t: [1, 3, 1], n_trials=1
-    )  # The trial result is the same as the above one.
-    assert {_trial_to_values(t) for t in study.best_trials} == {(1, 3, 1)}
-    assert len(study.best_trials) == 2
+    assert {tuple(t.values) for t in study.best_trials} == {(1, 1), (2, 2)}
 
 
 def test_wrong_n_objectives() -> None:
